Getting game by id from api
